Now able to calculate overall total
Clicking on the -/+ and trash buttons can now correctly calculate the overall total. But text input of quantities still not supported.
Also, if you decrease a quantity to 0 and then increase it back up to 1 again, the innerHTML of the price doesn't change to reflect that the first time. (The overall total is still calculated correctly.)
Code needs a lot of tidying up too!